refactor : 기존 필터링버튼을 FilterButtons, FilterOptionList 두개의 컴포넌트로 분리하여 재사용가능하게 변경(#7)
